[GAME][GRAPHICS] Pull Game Loop and Configurations to new Classes.
Game Loop is in its own class now. This currently has the screen implementation access inside of it. 
The configurations class is for storing configurations. 
The game Loop is to read from this to provide parameters for instantiating objects and classes. 
Theres a template function for passing the any classes to full classes. 
Screen implementation has been slightly modified so that we can store the screen class as a component of the gameloop class. 
This may be changed in the future.