[GAME] Added Physics Component empties.
Physics components are to maintain the cartesian polygons of the characters. 
[GAME] Added physics component to characters

Physics components are empty. 
Square Physics components are empty. 
They should generate an object in cartesian, to be converted to screen coords. 
They should be made to output some type of renderable square object to the square graphics component.